Initial command line/TCP server script
This fairly simple script leverages the Node API provided by CyberChef
to allow performing operations on local filesystem objects without
needing to copy them into and out of a browser window.

It requires a recipe to be specified at runtime as a JSON file.

Also includes a simple TCP server for use by other local processes. This
server is hard-coded to use localhost in order to discourage production
use of the script.