Update index.html
Added some text to help end users. It's a small but significant change to the UI that I think will make a big difference.

I have had CyberChef bookmarked for years and *every single time* I go to it I can't remember how it works. I click on Operations and click on this and click on that and for the life of me can't get it going. Many times I left without getting any results. 

I'm not a luddite, I'm a bit of a power user and even a web developer. IMHO, the small change I've made will help countless users. In fact, if implemented, I suggest you make a note in the website's analytics of the date you added this change; I think you'll see the number of users, and return users, start to go up significantly from that date.

P.S. You will probably also have to update the CSS:

    .title { white-space: nowrap; }