Add dependency on 'sass'
Without a dependency on sass it fails to build with the attached error on my
linux 64-bit system. This adds a devDependency on `sass` with the current
latest version.

The issue can also be solved by running `npm install sass`.

I'll be honest I don't fully understand the nuances of why this is needed, I
just tried it thanks to [this stackoverflow answer](https://stackoverflow.com/a/65796287).